+++ libraries/BLE/src/BLECharacteristic.h
@@ -62,7 +62,7 @@ public:
 	BLEUUID        getUUID();
 	std::string    getValue();
 	uint8_t*       getData();
-	uint8_t        getLength();
+	size_t         getLength();
 
 	void indicate();
 	void notify(bool is_notification = true);
